+++ tools/diagraphica/colorchooser.cpp
@@ -54,7 +54,7 @@ void ColorChooser::setPoints(
 // -----------------------------
 {
     positionsX.clear();
-    for ( int i = 0; i < hue.size(); ++i )
+    for ( size_t i = 0; i < hue.size(); ++i )
         positionsX.push_back( (2.0*hue[i])-1.0 );
     
     positionsY = y;
@@ -496,12 +496,12 @@ void ColorChooser::handleHits( const vector< int > &ids )
 {
     if ( mouseSide == MSE_SIDE_LFT )
     {
-        if ( 0 <= ids[0] && ids[0] < positionsX.size() )
+        if ( 0 <= ids[0] && static_cast <size_t> (ids[0]) < positionsX.size() )
             dragIdx = ids[0];
     }
     else if ( mouseSide == MSE_SIDE_RGT )
     {
-        if ( 0 <= ids[0] && ids[0] < positionsX.size() )
+        if ( 0 <= ids[0] && static_cast <size_t> (ids[0]) < positionsX.size() )
         {
             /*
             positionsX.erase( positionsX.begin()+ids[0] );
@@ -518,7 +518,7 @@ void ColorChooser::handleHits( const vector< int > &ids )
 void ColorChooser::handleDrag()
 // ----------------------------
 {
-    if ( 0 <= dragIdx && dragIdx < positionsX.size() )
+    if ( 0 <= dragIdx && static_cast <size_t> (dragIdx) < positionsX.size() )
     {
         double w, h;
         double xLft, xRgt;
